+++ src/renderer/src/components/project/ProjectView.tsx
@@ -326,6 +326,15 @@ export const ProjectView = ({ project, isActive = false }: Props) => {
     promptFieldRef.current?.focus();
   };
 
+  const onSubmitted = () => {
+    if (question) {
+      if (question.answerFunction) {
+        question.answerFunction('n');
+      }
+      setQuestion(null);
+    }
+  };
+
   const showFileDialog = (readOnly: boolean) => {
     setAddFileDialogOptions({
       readOnly,
@@ -484,6 +493,7 @@ export const ProjectView = ({ project, isActive = false }: Props) => {
             runTests={runTests}
             openModelSelector={() => projectTopBarRef.current?.openMainModelSelector()}
             disabled={!modelsData}
+            onSubmitted={onSubmitted}
           />
         </div>
       </div>